requirements.txt
streamlit
anthropic
python-dotenv
pypdf
snowflake-connector-python